requirements.txt
tensorflow
tensorflow-text
tf-models-official
pydot
pydotplus
graphviz